Change Terminal to use crossterm commands, refactor welcome message
I felt that it was getting to be a pain to write a separate terminal
function for just calling a command from crossterm. In addition, it
defeated the purpose of commands in the first place, and led to a lot of
execute calls instead of queueing and then flushing a list of commands.
I added the 'q' method to Terminal to allow the queueing of commands and
the chaining thereof.