Remove previous implementations of centering text
I realized that centering text is pretty much only needed for drawing
the welcome screen, and so does not need to be generalizable.
Furthermore, I intend to reimplement specific, welcome screen centering
for Window. There should be no case to ever center text directly on the
terminal itself -- use a full-terminal window instead.